refactor(exporter): move opensearch into a deeper package
The elastic and opensearch exporters were both using the same package structure. This caused issues during the CI.

I've changed the package from `io.camunda.zeebe.exporter` to `io.camunda.zeebe.exporter.opensearch`
