+++ exporters/opensearch-exporter/src/main/java/io/camunda/zeebe/exporter/opensearch/OpensearchExporterException.java
@@ -5,9 +5,15 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.exporter.dto;
+package io.camunda.zeebe.exporter.opensearch;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+public class OpensearchExporterException extends RuntimeException {
 
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record PutIndexTemplateResponse(boolean acknowledged) {}
+  public OpensearchExporterException(final String message) {
+    super(message);
+  }
+
+  public OpensearchExporterException(final String message, final Throwable cause) {
+    super(message, cause);
+  }
+}
