+++ exporters/opensearch-exporter/src/main/java/io/camunda/zeebe/exporter/opensearch/dto/Template.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.exporter.dto;
+package io.camunda.zeebe.exporter.opensearch.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
